embedded_dma

Trait ReadTarget

pub unsafe trait ReadTarget {
    type Word: Word;

    // Provided method
    fn as_read_buffer(&self) -> (*const Self::Word, usize) { ... }
}

Trait for Deref targets used by the blanket DmaReadBuffer impl.

This trait exists solely to work around https://github.com/rust-lang/rust/issues/20400.

Safety

as_read_buffer must adhere to the safety requirements documented for ReadBuffer::read_buffer.

Required Associated Types

type Word: Word

Provided Methods

fn as_read_buffer(&self) -> (*const Self::Word, usize)