Trait embedded_dma::ReadTarget

pub unsafe trait ReadTarget {
    type Word: Word;
    fn as_read_buffer(&self) -> (*const Self::Word, usize) { ... }
}

Trait for Deref targets used by the blanket DmaReadBuffer impl.

This trait exists solely to work around https://github.com/rust-lang/rust/issues/20400.

Safety

as_read_buffer must adhere to the safety requirements documented for ReadBuffer::read_buffer.

Associated Types

type Word: Word

Provided methods

fn as_read_buffer(&self) -> (*const Self::Word, usize)